+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Version 2.0.0
 - Added plugins directory to the views path.  This makes it possible for
   plugins to render their own ejs views or pages in their own folders by
   using paths of the format "pluginName/path/to/page-or-view".
+- Added instanceName entry to the config passed to instance plugins.
 - Replaced the per instance copy of the shared Factorio mods with
   symlinks.  On Windows hard links are used instead due to the
   privileges requirements of symlinks.
@@ -34,6 +35,12 @@ Version 2.0.0
 - The game port, RCON port and RCON password instance config entries are are
   now null by default, indicating that a random one will be generated every
   time the instance is started.
+- Removed the broken client download command as this would download
+  pre-releases.
+- Moved the item database and HTTP definitions for /api/place, /api/remove,
+  /api/inventory, /api/inventoryAsObject and the web interface view for the
+  storage page into the clusterioMod plugin.  If you disable this plugin
+  then these things will not be available.
 
 ### Breaking Changes